rework logger imports
