+++ packages/core/src/logger/no-op-logger.ts
@@ -1,0 +1,8 @@
+import type { BaseLogger } from './base-logger';
+
+export const noopLogger = {
+  debug: () => {},
+  info: () => {},
+  warn: () => {},
+  error: () => {},
+} as BaseLogger;
