+++ packages/core/src/logger/types.ts
@@ -1,0 +1,36 @@
+import type { Run } from '../run/types';
+
+export interface BaseLogMessage extends Run {
+  msg: string;
+  level: number;
+  time: Date;
+  pid: number;
+  hostname: string;
+  name: string;
+}
+
+export const LogLevel = {
+  DEBUG: 'debug',
+  INFO: 'info',
+  WARN: 'warn',
+  ERROR: 'error',
+  NONE: 'silent',
+} as const;
+
+export type LogLevel = (typeof LogLevel)[keyof typeof LogLevel];
+
+export const RegisteredLogger = {
+  AGENT: 'AGENT',
+  WORKFLOW: 'WORKFLOW',
+  LLM: 'LLM',
+  TTS: 'TTS',
+  VOICE: 'VOICE',
+  VECTOR: 'VECTOR',
+  BUNDLER: 'BUNDLER',
+  DEPLOYER: 'DEPLOYER',
+  MEMORY: 'MEMORY',
+  STORAGE: 'STORAGE',
+  EMBEDDINGS: 'EMBEDDINGS',
+} as const;
+
+export type RegisteredLogger = (typeof RegisteredLogger)[keyof typeof RegisteredLogger];
