+++ packages/core/tsup.config.ts
@@ -44,6 +44,10 @@ export default defineConfig({
     'src/utils.ts',
     '!src/action/index.ts',
     'src/*/index.ts',
+    'src/logger/base-logger.ts',
+    'src/logger/logger.ts',
+    'src/logger/multi-logger.ts',
+    'src/logger/transport.ts',
     'src/storage/libsql/index.ts',
     'src/vector/libsql/index.ts',
     'src/vector/filter/index.ts',
